requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
pydantic==2.5.0
python-multipart==0.0.6
dash==2.14.2
plotly==5.18.0
pandas==2.1.3
requests==2.31.0